[plugin] fix compile dependency of test client
`GraphQLGenerateTestClientTask` was extending `GraphQLGenerateClientTask` which meant that we were configuring all task of type `GraphQLGenerateClientTask` we were also configuring test client. Moved the common client generation logic to a new `AbstractGenerateClientTask` which is used by both client generation tasks.
