--- examples/client/gradle-client/settings.gradle.kts
@@ -1,1 +1,0 @@
-rootProject.name = "graphql-kotlin-gradle-client-example"
